# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MemberView": {
        "properties": {
          "role": {
            "default": "owner",
            "title": "Role",
            "type": "string"
          },
          "user_ref": {
            "title": "User Ref",
            "type": "string"
          }
        },
        "required": [
          "user_ref"
        ],
        "title": "MemberView",
        "type": "object"
      },
...
      "ReadTokenCreate": {
        "properties": {
          "label": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Label",
            "type": "string"
          },
          "role": {
            "enum": [
              "viewer",
              "responder",
              "admin",
              "auditor"
            ],
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "label",
          "role"
        ],
        "title": "ReadTokenCreate",
        "type": "object"
      },
...
      "SettingsUpdate": {
        "properties": {
          "capture_payloads": {
            "title": "Capture Payloads",
            "type": "boolean"
          },
          "retention_days": {
            "maximum": 3650.0,
            "minimum": 1.0,
            "title": "Retention Days",
            "type": "integer"
          },
          "tenant_name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Tenant Name",
            "type": "string"
          }
        },
        "required": [
          "tenant_name",
          "retention_days",
          "capture_payloads"
        ],
        "title": "SettingsUpdate",
        "type": "object"
      },
      "SettingsView": {
        "properties": {
          "capture_payloads": {
            "title": "Capture Payloads",
            "type": "boolean"
          },
          "compliance_mode": {
            "title": "Compliance Mode",
            "type": "boolean"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "members": {
            "items": {
              "$ref": "#/components/schemas/MemberView"
            },
            "title": "Members",
            "type": "array"
          },
          "plan_key": {
            "title": "Plan Key",
            "type": "string"
          },
          "redaction_rules": {
            "items": {
              "type": "string"
            },
            "title": "Redaction Rules",
            "type": "array"
          },
          "retention_days": {
            "title": "Retention Days",
            "type": "integer"
          },
          "retention_floor_days": {
            "title": "Retention Floor Days",
            "type": "integer"
          },
          "tenant_id": {
            "title": "Tenant Id",
            "type": "string"
          },
          "tenant_name": {
            "title": "Tenant Name",
            "type": "string"
          },
          "tokens": {
            "items": {
              "$ref": "#/components/schemas/TokenView"
            },
            "title": "Tokens",
            "type": "array"
          },
          "usage": {
            "$ref": "#/components/schemas/UsageView"
          }
        },
        "required": [
          "tenant_id",
          "tenant_name",
          "created_at",
          "members",
          "tokens",
          "retention_days",
          "capture_payloads",
          "compliance_mode",
          "retention_floor_days",
          "redaction_rules",
          "plan_key",
          "usage"
        ],
        "title": "SettingsView",
        "type": "object"
      },
...
      "TokenCreate": {
        "properties": {
          "kind": {
            "enum": [
              "recording",
              "local_scan"
            ],
            "title": "Kind",
            "type": "string"
          },
          "label": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Label",
            "type": "string"
          },
          "operator_user_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operator User Ref"
          }
        },
        "required": [
          "kind",
          "label"
        ],
        "title": "TokenCreate",
        "type": "object"
      },
      "TokenCreated": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "token": {
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "label",
          "token"
        ],
        "title": "TokenCreated",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Leaflyst API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/evidence/tenant": {
      "get": {
        "description": "Stream the complete tenant chain through its latest immutable anchor.",
        "operationId": "export_tenant_evidence_v1_evidence_tenant_get",
        "parameters": [
          {
            "in": "query",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "title": "Tenant Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-abx-admin-key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Abx-Admin-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Tenant Evidence"
      }
    },
...
    "/v1/settings": {
      "get": {
        "operationId": "get_settings_v1_settings_get",
        "parameters": [
          {
            "in": "query",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "title": "Tenant Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-abx-admin-key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Abx-Admin-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Settings"
      },
      "put": {
        "operationId": "update_settings_v1_settings_put",
        "parameters": [
          {
            "in": "query",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "title": "Tenant Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-abx-admin-key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Abx-Admin-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Settings"
      }
    },
    "/v1/settings/read-tokens": {
      "post": {
        "description": "Mint a tenant-BOUND read token with an explicit role.\n\nReplaces reaching for the shared operator key. The token carries its tenant,\nso it cannot be pointed at another one, and its role, so an auditor handed\none cannot change anything.",
        "operationId": "create_read_token_v1_settings_read_tokens_post",
        "parameters": [
          {
            "in": "query",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "title": "Tenant Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-abx-admin-key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Abx-Admin-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReadTokenCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Read Token"
      }
    },
    "/v1/settings/read-tokens/{token_id}/revoke": {
      "post": {
        "operationId": "revoke_read_token_v1_settings_read_tokens__token_id__revoke_post",
        "parameters": [
          {
            "in": "path",
            "name": "token_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Token Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "title": "Tenant Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-abx-admin-key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Abx-Admin-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Revoke Read Token V1 Settings Read Tokens  Token Id  Revoke Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke Read Token"
      }
    },
    "/v1/settings/tokens": {
      "post": {
        "operationId": "create_token_v1_settings_tokens_post",
        "parameters": [
          {
            "in": "query",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "title": "Tenant Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-abx-admin-key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Abx-Admin-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Token"
      }
    },
    "/v1/settings/tokens/{kind}/{token_id}/revoke": {
      "post": {
        "operationId": "revoke_token_v1_settings_tokens__kind___token_id__revoke_post",
        "parameters": [
          {
            "in": "path",
            "name": "kind",
            "required": true,
            "schema": {
              "enum": [
                "recording",
                "local_scan"
              ],
              "title": "Kind",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "token_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Token Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "title": "Tenant Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-abx-admin-key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Abx-Admin-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Revoke Token V1 Settings Tokens  Kind   Token Id  Revoke Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke Token"
      }
    }
  }
}